move: step1에 작성한 코드를 step1 패키지로 이동
